Fix warnings

+++ lib/datastruct/patricia.c
@@ -322,7 +322,7 @@ patricia_insert(PATRICIA * P, const uint8_t * key, size_t keylen, void * rec)
 				slen = MAXSLEN;
 
 			/* Create the node or error out. */
-			if ((pnew = node_alloc(slen, key)) == NULL)
+			if ((pnew = node_alloc((uint8_t)slen, key)) == NULL)
 				return (-1);
 
 			/* Add the new node into the tree. */
@@ -340,7 +340,7 @@ patricia_insert(PATRICIA * P, const uint8_t * key, size_t keylen, void * rec)
 			 */
 			/* Create the lower of the new nodes. */
 			slen = (*np)->slen - mlen;
-			pnew2 = node_dup(*np, slen, (*np)->s + mlen);
+			pnew2 = node_dup(*np, (uint8_t)slen, (*np)->s + mlen);
 			if (pnew2 == NULL)
 				return (-1);
